pyproject.toml
[tool.poetry]
name = "coffee-roaster-market"
version = "0.1.0"
description = "Django based - coffee roaster web market"
authors = ["Aleksander Pawlak <qaleksander.pawlak1996@gmail.com>", "Paweł Kucharski <pawel.kucharski@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-rest-framework = "^0.1.0"
django-environ = "^0.4.5"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"
django-templated-email = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.1.0"
black = "^20.8b1"
flake8 = "^3.8.4"
django-stubs = "^1.8.0"
mypy = "^0.812"
factory-boy = "^3.2.0"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
